Request audio permission when necessary in the Android app
Previous to this commit, the permission was asked when the app first
started. If the user tapped 'Deny' there, then there was no option to
undo that choice.

The permission is now always asked when it is requested by the webpage, and if
it hasn't been granted before, instead of when first opening the app.