Add pilot challenge team info to the constants file (#2266)
As we start hooking the front- and back-end together, it's important to
have stricter checks in place that we're sending data in the correct
format. This file should help with that.

Eventually, all the team string literals we use in our challenge logic
should be replaced with values from this file.

We decided to add a front-end specific constants file, since our webpack
config is inconsistent between /common and /web so broke on the `require`
calls used for logo URLs.

Note: the Mozilla and IBM svgs that I added are placeholders until we
get updated assets from design.