Clear some cobwebs from our page layout (#2863)
Prompted by a surprising layout breakage on our new error pages, I found some oddities in our page layout. This is my attempt to simplify a few years of uh, iterative development.

I did a bunch of detective work and testing while going through this, which I've tried to document below. But please let me know if you have any questions or concerns!

I've got a big change to follow this one to get rid of the `#scroller` and `#scrollee` containers, but that deserves its own commit and explanation. This is some comparatively simple prep work.

Here are the changes, explained:

---

- Remove an unused `.focus` style.

Added in 2017 (pre e416be58b5aa), this class was used to add custom Firefox Focus scroll behaviour. The class was removed in ce541c8c2867, but I accidentally left a custom rule in place (unused since ce541c8c2867).

- Remove some unused / duplicate `hasScrolled` state from `app.tsx`.

Honestly I didn't dive deep into this one because it's obviously not being used. The same `hasScrolled` behaviour is duplicated in `layout.tsx` so I assume it was just moved-but-not-deleted somewhere along the way.

- Remove an unused `.hidden` class (and associated `hasScrolledDown` logic).

It looks like there were plans to make the header hide itself on scroll, but that never happened. This was added in e4c48c8f01a1 and never used. Better solutions exist in 2020, so rather than keep this around we should get rid of it.

- Remove an unused `.hidden` style.

Confusing! There _was_ a `.hidden` class (see above), but not within `.faq-hero`. This was added in 762110ae30b3 and apparently never used.

---

And that's it. Just in time for some real QA before release! Which is great, because this is gonna need it :)