Dupe encoding (#3261)
* Fix weird duplicate encoding issue

Sometimes (maybe when the server is under load and not responsive)
the client sends multiple copies of a file to the server, and diff
pods all try to encode the file at the same time, which both
is an inefficient use of resources. The errors from these also makes
it difficult to diagnose real FFmpeg issues from weird collision ones.
This PR checks if a file exists on S3 before beginning the encoding
process, and if yes, skips it. It incidentally also:

* Fixes issues where throw errors inside a request don't terminate
  the rest of the function
* Ensures tasks for after S3 upload actually happens after the entire
  stream is encoded

* Send encoding errors to Sentry instead of just logging