Sort languages in the Data Sets > Languages dropdown, accounting for localization. (#3301)
* #3295: Add sort logic for locales.

This commit adds localized sort logic for the language dropdown. It
gets the language name for the locales and then uses localized
comparison so that non-Latin and accented characters are sorted
correctly.

* #3295: Remove unnecessary empty line additions.

* #3295: Add `useSortedLocales` hook; Specify `getString` type

This commit sets up a new hook that can be used to fetch a list of
sorted locales. The list is sorted based on the name of each locale,
localized for the current locale of the client.

* #3295: Leverage `useSortedLocales` hook for Datasets and LanguageSelect

* #3295: Insert spaces to make import style consistent